webhook-deployment/pyproject.toml
[project]
name = "webhook-deployment"
version = "0.1.0"
description = "GitHub webhook receiver for auto-deployment of Matrix PDF Summarizer Bot"
authors = [{ name = "Carlos Gomes", email = "carlos@example.com" }]
requires-python = ">=3.9"

dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]